html email manager is now the main email sending task
